refactor(main.cpp): Split into AccessPointManager & RobotTaskManager
