+++ src/RobotTaskManager.cpp
@@ -1,0 +1,125 @@
+#include "RobotTaskManager.h"
+
+namespace Robot {
+	void RobotTaskManager::begin() {
+		robot.begin();
+	}
+
+	void RobotTaskManager::RobotTask(void* pvParameters) {
+		auto* self = static_cast<RobotTaskManager*>(pvParameters);
+		self->robot.playStartupTone();
+		self->robotTaskLoop();
+	}
+
+	void RobotTaskManager::startTask() {
+		// Create a queue for robot commands (queue length of 10).
+		robotQueue = xQueueCreate(10, sizeof(PalookaNetwork::CommandData));
+
+		// Create the hardware control task pinned to core 1.
+		xTaskCreatePinnedToCore(
+			RobotTaskManager::RobotTask,
+			"RobotTask",	// Task name
+			4096,			// Stack size
+			this,			// Parameter
+			2,				// Priority
+			nullptr,		// Task handle
+			1				// Pin to core 1
+		);
+	}
+
+	void RobotTaskManager::robotTaskLoop()
+	{
+		static const unsigned long batteryUpdateInterval = 5000; // 5 seconds
+		static unsigned long lastBatteryUpdate = 0; // Last battery update time
+		static const unsigned long ledToggleInterval = 1000; // 1 second
+		static unsigned long lastLedToggle = 0; // Last time LED was toggled
+		while(true)
+		{
+			unsigned long currentMillis = millis();
+			if((currentMillis - lastBatteryUpdate) >= batteryUpdateInterval)
+			{
+				sendBatteryUpdate();
+				lastBatteryUpdate = currentMillis;
+			}
+			if((currentMillis - lastLedToggle) >= ledToggleInterval)
+			{
+				robot.toggleLed();
+				lastLedToggle = currentMillis;
+			}
+
+			handleWebsocketCommands(); // Non-blocking
+
+			// Minor yield for other tasks to execute
+			vTaskDelay(pdMS_TO_TICKS(1));
+		}
+	}
+
+	void RobotTaskManager::handleWebsocketCommands()
+	{
+		PalookaNetwork::CommandData cmdData;
+		// Wait for a command with a 10ms timeout
+		if(xQueueReceive(robotQueue, &cmdData, pdMS_TO_TICKS(10)) != pdPASS) { return; }
+
+		// Process slider control data
+		if(cmdData.hasSlider)
+		{
+			char robotLimb = cmdData.sliderName[0];
+			Serial.print("Limb: ");
+			Serial.print(robotLimb);
+			Serial.print(", Value: ");
+			Serial.println(cmdData.value);
+
+			handleRobotSliderCommand(robotLimb, cmdData.value);
+		}
+		// Process joystick control data
+		else if(cmdData.hasJoystick)
+		{
+			Serial.print("Joystick X: ");
+			Serial.print(cmdData.x);
+			Serial.print(", Y: ");
+			Serial.println(cmdData.y);
+
+			robot.move(cmdData.x, cmdData.y);
+		}
+		// Process flip command
+		else if(cmdData.hasFlip)
+		{
+			robot.flip();
+		}
+		else
+		{
+			Serial.println("Unknown command structure.");
+		}
+	}
+
+	void RobotTaskManager::handleRobotSliderCommand(const char robotLimb, const int value)
+	{
+		switch(robotLimb)
+		{
+			case 'R': case 'r': // Right motor
+				robot.moveRightWheel(value);
+				break;
+			case 'L': case 'l': // Left motor
+				robot.moveLeftWheel(value);
+				break;
+			case 'F': case 'f': // Flipper arm
+				robot.moveFlipper(value);
+				break;
+			default:
+				Serial.println("Unknown robot limb JSON supplied.");
+				break;
+		}
+	}
+
+	void RobotTaskManager::sendBatteryUpdate()
+	{
+		int batteryLevel = robot.getBatteryPercentage();
+		StaticJsonDocument<100> batteryDoc;
+		batteryDoc["battery"] = batteryLevel;
+		String batteryJson;
+		serializeJson(batteryDoc, batteryJson);
+
+		// Send the JSON string to all connected clients
+		this->apManager.sendWebSocketMessage(batteryJson);
+	}
+}
